requirements.txt
Pillow==10.0.0
scikit_learn==1.2.2
streamlit==1.12.0
